Move PSBT ser/deser stuff to module root
Code move only, no logic changes.
