+++ bitcoin/src/psbt/mod.rs
@@ -14,8 +14,11 @@ mod map;
 pub mod raw;
 pub mod serialize;
 
+use self::map::Map;
 use crate::bip32::{KeySource, Xpub};
-use crate::prelude::{BTreeMap, Vec};
+use crate::consensus::encode::Decodable;
+use crate::io::Write;
+use crate::prelude::{BTreeMap, DisplayHex, Vec};
 use crate::transaction::Transaction;
 
 #[rustfmt::skip]                // Keep public re-exports separate.
@@ -63,6 +66,94 @@ impl Psbt {
         }
 
         Ok(())
+    }
+
+    /// Serialize a value as bytes in hex.
+    pub fn serialize_hex(&self) -> String { self.serialize().to_lower_hex_string() }
+
+    /// Serialize as raw binary data
+    pub fn serialize(&self) -> Vec<u8> {
+        let mut buf: Vec<u8> = Vec::new();
+        self.serialize_to_writer(&mut buf).expect("Writing to Vec can't fail");
+        buf
+    }
+
+    /// Serialize the PSBT into a writer.
+    pub fn serialize_to_writer(&self, w: &mut impl Write) -> io::Result<usize> {
+        let mut written_len = 0;
+
+        fn write_all(w: &mut impl Write, data: &[u8]) -> io::Result<usize> {
+            w.write_all(data).map(|_| data.len())
+        }
+
+        // magic
+        written_len += write_all(w, b"psbt")?;
+        // separator
+        written_len += write_all(w, &[0xff])?;
+
+        written_len += write_all(w, &self.serialize_map())?;
+
+        for i in &self.inputs {
+            written_len += write_all(w, &i.serialize_map())?;
+        }
+
+        for i in &self.outputs {
+            written_len += write_all(w, &i.serialize_map())?;
+        }
+
+        Ok(written_len)
+    }
+
+    /// Deserialize a value from raw binary data.
+    pub fn deserialize(mut bytes: &[u8]) -> Result<Self, Error> {
+        Self::deserialize_from_reader(&mut bytes)
+    }
+
+    /// Deserialize a value from raw binary data read from a `BufRead` object.
+    pub fn deserialize_from_reader<R: io::BufRead>(r: &mut R) -> Result<Self, Error> {
+        const MAGIC_BYTES: &[u8] = b"psbt";
+
+        let magic: [u8; 4] = Decodable::consensus_decode(r)?;
+        if magic != MAGIC_BYTES {
+            return Err(Error::InvalidMagic);
+        }
+
+        const PSBT_SERPARATOR: u8 = 0xff_u8;
+        let separator: u8 = Decodable::consensus_decode(r)?;
+        if separator != PSBT_SERPARATOR {
+            return Err(Error::InvalidSeparator);
+        }
+
+        let mut global = Psbt::decode_global(r)?;
+        global.unsigned_tx_checks()?;
+
+        let inputs: Vec<Input> = {
+            let inputs_len: usize = (global.unsigned_tx.input).len();
+
+            let mut inputs: Vec<Input> = Vec::with_capacity(inputs_len);
+
+            for _ in 0..inputs_len {
+                inputs.push(Input::decode(r)?);
+            }
+
+            inputs
+        };
+
+        let outputs: Vec<Output> = {
+            let outputs_len: usize = (global.unsigned_tx.output).len();
+
+            let mut outputs: Vec<Output> = Vec::with_capacity(outputs_len);
+
+            for _ in 0..outputs_len {
+                outputs.push(Output::decode(r)?);
+            }
+
+            outputs
+        };
+
+        global.inputs = inputs;
+        global.outputs = outputs;
+        Ok(global)
     }
 }
