+++ bitcoin/src/psbt/serialize.rs
@@ -8,14 +8,13 @@
 use hashes::{hash160, ripemd160, sha256, sha256d};
 use secp256k1::XOnlyPublicKey;
 
-use super::map::{Input, Map, Output, PsbtSighashType};
+use super::map::PsbtSighashType;
 use crate::bip32::{ChildNumber, Fingerprint, KeySource};
 use crate::consensus::encode::{self, deserialize_partial, serialize, Decodable, Encodable};
 use crate::crypto::key::PublicKey;
 use crate::crypto::{ecdsa, taproot};
-use crate::io::Write;
-use crate::prelude::{DisplayHex, String, Vec};
-use crate::psbt::{Error, Psbt};
+use crate::prelude::Vec;
+use crate::psbt::Error;
 use crate::script::ScriptBuf;
 use crate::taproot::{
     ControlBlock, LeafVersion, TapLeafHash, TapNodeHash, TapTree, TaprootBuilder,
@@ -36,95 +35,6 @@ pub(crate) trait Deserialize: Sized {
     fn deserialize(bytes: &[u8]) -> Result<Self, Error>;
 }
 
-impl Psbt {
-    /// Serialize a value as bytes in hex.
-    pub fn serialize_hex(&self) -> String { self.serialize().to_lower_hex_string() }
-
-    /// Serialize as raw binary data
-    pub fn serialize(&self) -> Vec<u8> {
-        let mut buf: Vec<u8> = Vec::new();
-        self.serialize_to_writer(&mut buf).expect("Writing to Vec can't fail");
-        buf
-    }
-
-    /// Serialize the PSBT into a writer.
-    pub fn serialize_to_writer(&self, w: &mut impl Write) -> io::Result<usize> {
-        let mut written_len = 0;
-
-        fn write_all(w: &mut impl Write, data: &[u8]) -> io::Result<usize> {
-            w.write_all(data).map(|_| data.len())
-        }
-
-        // magic
-        written_len += write_all(w, b"psbt")?;
-        // separator
-        written_len += write_all(w, &[0xff])?;
-
-        written_len += write_all(w, &self.serialize_map())?;
-
-        for i in &self.inputs {
-            written_len += write_all(w, &i.serialize_map())?;
-        }
-
-        for i in &self.outputs {
-            written_len += write_all(w, &i.serialize_map())?;
-        }
-
-        Ok(written_len)
-    }
-
-    /// Deserialize a value from raw binary data.
-    pub fn deserialize(mut bytes: &[u8]) -> Result<Self, Error> {
-        Self::deserialize_from_reader(&mut bytes)
-    }
-
-    /// Deserialize a value from raw binary data read from a `BufRead` object.
-    pub fn deserialize_from_reader<R: io::BufRead>(r: &mut R) -> Result<Self, Error> {
-        const MAGIC_BYTES: &[u8] = b"psbt";
-
-        let magic: [u8; 4] = Decodable::consensus_decode(r)?;
-        if magic != MAGIC_BYTES {
-            return Err(Error::InvalidMagic);
-        }
-
-        const PSBT_SERPARATOR: u8 = 0xff_u8;
-        let separator: u8 = Decodable::consensus_decode(r)?;
-        if separator != PSBT_SERPARATOR {
-            return Err(Error::InvalidSeparator);
-        }
-
-        let mut global = Psbt::decode_global(r)?;
-        global.unsigned_tx_checks()?;
-
-        let inputs: Vec<Input> = {
-            let inputs_len: usize = (global.unsigned_tx.input).len();
-
-            let mut inputs: Vec<Input> = Vec::with_capacity(inputs_len);
-
-            for _ in 0..inputs_len {
-                inputs.push(Input::decode(r)?);
-            }
-
-            inputs
-        };
-
-        let outputs: Vec<Output> = {
-            let outputs_len: usize = (global.unsigned_tx.output).len();
-
-            let mut outputs: Vec<Output> = Vec::with_capacity(outputs_len);
-
-            for _ in 0..outputs_len {
-                outputs.push(Output::decode(r)?);
-            }
-
-            outputs
-        };
-
-        global.inputs = inputs;
-        global.outputs = outputs;
-        Ok(global)
-    }
-}
 impl_psbt_de_serialize!(Transaction);
 impl_psbt_de_serialize!(TxOut);
 impl_psbt_de_serialize!(Witness);
@@ -400,6 +310,7 @@ fn key_source_len(key_source: &KeySource) -> usize { 4 + 4 * (key_source.1).as_r
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::psbt::Psbt;
     use crate::script::ScriptBufExt as _;
 
     // Composes tree matching a given depth map, filled with dumb script leafs,
